introduce Storage interface and simple implementation
It makes sense to allow to configure storage type depending on wanted
features and speed/memory requirements.
Using SPLFixedArrayStorage is fixed for single long-running process, but
is useless for processes that die fast (when serving HTTP requests for
example).
So there should be a way to store data between requests. New storages can
be implemented for Redis for instance.